pipeline/requirements.txt
scrapy
scrapyd-client
boto
demjson
python-dateutil
